IOS-3682 Separate to DerivationProviders
